Practicament llest per entregar

+++ src/main/java/com/liceu/sromerom/sakilaExample/repos/RentalRepo.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface RentalRepo {
     List<Rental> findAll();
-    int rent(String rentaldate, long inventoryid, long customerid, long staffid);
+    Long rent(String rentaldate, long inventoryid, long customerid, long staffid);
     int returnDVD(String returndate, long rentalid);
     Rental getRentalByInventoryAndCustomer(long inventoryid, long customerid);
 }
